developer.wordpress.dev: Don't run the parser on provision.
It takes a very long time, which makes for a bad user experience when someone is setting it up for the first time,
especially after the normal Vagrant provisioning has already taken a 20+ minutes. The database already has a
complete set of records for 3.9, and this can be updated periodically, and of course they can manually run the
parser to update it themselves.