WordPressorg.dev: Prune unnecessary content to reduce database size.
This mainly focuses on develop.wordpressorg.dev and translate.wordpressorg.dev. In both cases there were tens to hundreds of thousands of rows of similar content. That added little benefit, but makes it much harder to review database merges.

This dropped the export size from 28MB to 5MB, while preserving enough sample content for contributors to work with. For developer.wordpressorg.dev, all of the content can be restored by re-running the parser.

See #28
See #50