Move configs to root folder to optimize provision speed
VVV preferentially checks `.vvv/` and `provision` subfolders in the root to save time. If it doesn't find what it's looking for though, it'll do a costly subfolder search.

At the moment the main example for keeping this subfolder search is this Meta Environment. In addition, it will always load all the Nginx configs and all the hosts files anyway because of how the provisioner works, even if the user only wants `wptv` provisioning for example.

`.vvv` is the most ideal folder name today, but `provision` was chosen because for its back-compat benefits.


The hardcoded folder names were also removed, so that the Meta Environment will still work if it's named differently.

See #9
See #125